renames and deletes
